+++ tests/test_sensor.py
@@ -2,8 +2,28 @@
 
 from unittest.mock import MagicMock, Mock
 
-from custom_components.winix.const import ATTR_AIR_QUALITY, ATTR_AIR_QVALUE
-from custom_components.winix.sensor import ICON, UNIT_OF_MEASUREMENT, WinixSensor
+from custom_components.winix.const import ATTR_AIR_QUALITY, ATTR_AIR_QVALUE, DOMAIN
+from custom_components.winix.sensor import (
+    ICON,
+    UNIT_OF_MEASUREMENT,
+    WinixSensor,
+    async_setup_platform,
+)
+
+from tests import build_fake_manager
+
+
+async def test_setup_platform():
+    """Test platform setup."""
+
+    manager = build_fake_manager(3)
+    hass = Mock()
+    hass.data = {DOMAIN: manager}
+
+    async_add_entities = Mock()
+    await async_setup_platform(hass, None, async_add_entities, None)
+    assert async_add_entities.called
+    assert len(async_add_entities.call_args[0][0]) == 3
 
 
 def test_sensor_construction():
